Moved from python-standalone to Django
The application is now a Django project. This means that it has a (very
basic at the moment) web interface. It still provides most of the
functionality except for giving verbose ranking information (such as the
circular jugg differences, although they are used internally) and
drawing the graph using graphviz.
Those features are however planned to be re-implemented.