Cosmetic improvements
Ranking table now shows unranked (new) teams at the bottom by default.
They can be hidden by clicking on a link beneath the table. The setting
is currently not persistent (implemented in JS only).
The user can now click on teams to mark them in the ranking as well as
in the games list. Clicking the selected team again unmarks them. This
is also purely JS, so not persistent across page reloads.