Colored graphs
Fixes #2
Graphs are now colored. Circles have one color each, the same color in
both the full graph and in the circles graph (to make them easy to spot
in the full graph). In the tie breaker graph, different colors are used.
Teams of a circle have one color and teams considered for the tie
breaking have a different color (but they all have the same).